Fixes, formats, etc.
