+++ demo/src/index.js
@@ -3,7 +3,7 @@ import { render } from 'react-dom'
 import { Wrapper, Pre, LineNo } from './styles'
 
 import Highlight, { defaultProps } from 'prism-react-renderer'
-import theme from 'prism-react-renderer/themes/oceanicNext'
+import theme from 'prism-react-renderer/src/themes/oceanicNext'
 
 const exampleCode = `
 (function someDemo() {
