+++ src/defaultProps.js
@@ -1,9 +1,9 @@
 // @flow
 
-import Prism from './vendor/prism'
-import theme from '../themes/duotoneDark'
+import Prism from "./vendor/prism";
+import theme from "./themes/duotoneDark";
 
-import type { PrismLib } from './types'
+import type { PrismLib } from "./types";
 
 const defaultProps = {
   // $FlowFixMe
