+++ src/themes/dracula.js
@@ -2,7 +2,7 @@
 // Original: https://github.com/dracula/visual-studio-code
 // Converted automatically using ./tools/themeFromVsCode
 
-/*:: import type { PrismTheme } from '../src/types' */
+/*:: import type { PrismTheme } from '../types' */
 
 var theme /*: PrismTheme */ = {
   plain: {
@@ -68,4 +68,4 @@ var theme /*: PrismTheme */ = {
   ]
 };
 
-module.exports = theme;
+export default theme;
