--- themes/duotoneDark.js
@@ -1,105 +1,0 @@
-// @flow
-// Duotone Dark
-// Author: Simurai, adapted from DuoTone themes for Atom (http://simurai.com/projects/2016/01/01/duotone-themes)
-// Conversion: Bram de Haan (http://atelierbram.github.io/Base2Tone-prism/output/prism/prism-base2tone-evening-dark.css)
-// Generated with Base16 Builder (https://github.com/base16-builder/base16-builder)
-
-/*:: import type { PrismTheme } from '../src/types' */
-
-var theme /*: PrismTheme */ = {
-  plain: {
-    backgroundColor: "#2a2734",
-    color: "#9a86fd"
-  },
-  styles: [
-    {
-      types: ["comment", "prolog", "doctype", "cdata", "punctuation"],
-      style: {
-        color: "#6c6783"
-      }
-    },
-    {
-      types: ["namespace"],
-      style: {
-        opacity: 0.7
-      }
-    },
-    {
-      types: ["tag", "operator", "number"],
-      style: {
-        color: "#e09142"
-      }
-    },
-    {
-      types: ["property", "function"],
-      style: {
-        color: "#9a86fd"
-      }
-    },
-    {
-      types: ["tag-id", "selector", "atrule-id"],
-      style: {
-        color: "#eeebff"
-      }
-    },
-    {
-      types: ["attr-name"],
-      style: {
-        color: "#c4b9fe"
-      }
-    },
-    {
-      types: [
-        "boolean",
-        "string",
-        "entity",
-        "url",
-        "attr-value",
-        "keyword",
-        "control",
-        "directive",
-        "unit",
-        "statement",
-        "regex",
-        "at-rule",
-        "placeholder",
-        "variable"
-      ],
-      style: {
-        color: "#ffcc99"
-      }
-    },
-    {
-      types: ["deleted"],
-      style: {
-        textDecorationLine: "line-through"
-      }
-    },
-    {
-      types: ["inserted"],
-      style: {
-        textDecorationLine: "underline"
-      }
-    },
-    {
-      types: ["italic"],
-      style: {
-        fontStyle: "italic"
-      }
-    },
-    {
-      types: ["important", "bold"],
-      style: {
-        fontWeight: "bold"
-      }
-    },
-    {
-      types: ["important"],
-      style: {
-        color: "#c4b9fe"
-      }
-    }
-  ]
-};
-
-module.exports = theme;
