--- themes/duotoneLight.js
@@ -1,109 +1,0 @@
-// @flow
-// Duotone Light
-// Author: Simurai, adapted from DuoTone themes for Atom (http://simurai.com/projects/2016/01/01/duotone-themes)
-// Conversion: Bram de Haan (http://atelierbram.github.io/Base2Tone-prism/output/prism/prism-base2tone-evening-dark.css)
-// Generated with Base16 Builder (https://github.com/base16-builder/base16-builder)
-
-/*:: import type { PrismTheme } from '../src/types' */
-
-var theme /*: PrismTheme */ = {
-  plain: {
-    backgroundColor: "#faf8f5",
-    color: "#728fcb"
-  },
-  styles: [
-    {
-      types: ["comment", "prolog", "doctype", "cdata", "punctuation"],
-      style: {
-        color: "#b6ad9a"
-      }
-    },
-    {
-      types: ["namespace"],
-      style: {
-        opacity: 0.7
-      }
-    },
-    {
-      types: ["tag", "operator", "number"],
-      style: {
-        color: "#063289"
-      }
-    },
-    {
-      types: ["property", "function"],
-      style: {
-        color: "#b29762"
-      }
-    },
-    {
-      types: ["tag-id", "selector", "atrule-id"],
-      style: {
-        color: "#2d2006"
-      }
-    },
-    {
-      types: ["attr-name"],
-      style: {
-        color: "#896724"
-      }
-    },
-    {
-      types: [
-        "boolean",
-        "string",
-        "entity",
-        "url",
-        "attr-value",
-        "keyword",
-        "control",
-        "directive",
-        "unit",
-        "statement",
-        "regex",
-        "at-rule"
-      ],
-      style: {
-        color: "#728fcb"
-      }
-    },
-    {
-      types: ["placeholder", "variable"],
-      style: {
-        color: "#93abdc"
-      }
-    },
-    {
-      types: ["deleted"],
-      style: {
-        textDecorationLine: "line-through"
-      }
-    },
-    {
-      types: ["inserted"],
-      style: {
-        textDecorationLine: "underline"
-      }
-    },
-    {
-      types: ["italic"],
-      style: {
-        fontStyle: "italic"
-      }
-    },
-    {
-      types: ["important", "bold"],
-      style: {
-        fontWeight: "bold"
-      }
-    },
-    {
-      types: ["important"],
-      style: {
-        color: "#896724"
-      }
-    }
-  ]
-};
-
-module.exports = theme;
